Move valid_integration to util
